changed depic "dependencies"
changed milkshake custom addons dependency in the depiction from "com.anemonetheming.anemone" to just "Anemone" as the dependecy section in the depiction is for display purposes only. the dependency section in the deb control file is where the package bundle should be listed so cydia can properly associate the dependent-upon package.